backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Django==3.2.25
django-cors-headers==3.10.1
djangorestframework==3.15.1
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pytz==2025.2
sqlparse==0.4.4
typing-extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.0.3
zipp==3.6.0
